fix(features): limit max features
