+++ api/app/settings/common.py
@@ -896,3 +896,4 @@ DEFAULT_DOMAIN = "app.flagsmith.com"
 
 MAX_SEGMENTS_ALLOWED = env.int("MAX_SEGMENTS_ALLOWED", 30)
 MAX_SEGMENT_OVERRIDE_ALLOWED = env.int("MAX_SEGMENT_OVERRIDE_ALLOWED", 100)
+MAX_FEATURES_ALLOWED = env.int("MAX_FEATURES_ALLOWED", 200)
